added a dependency for proposed code of net.minecraft.server.PlayerChunkMap
The dependency is the static variable at the very top which allows the storage of the hashes as the other class is constantly being reopened within a loop of on of its parents which cannot be accessed within the bukkit api since mojang controls the class putting net.minecraft.server.PlayerChunkMap in a loop. So it has been placed in here the parent of the looping thread.