Fixed empty chunks sending to client and same chunk sending twice meaning more fps
In this code chunks are now cached as hashes in a list and if the same one appears twice for the same chunk and user then the send request for that chunk will be canceled. In addition if the chunk is an air chunk it will also be canceled as clients by default will generate the chunks as air there by boosting the fps of skyblock and skyland servers by a lot. Keep not that with these implementations it increases the use of resources by a touch but tps (Ticks per second) remains at the standard level of 20 and fps boosts with an additional 10 in the worst case scenario or the best case scenario it will add an additional 30 or 40 fps.